refactor: update imports to utilize shouldDisplayChainLogo utility
- Added the shouldDisplayChainLogo utility to multiple components for improved logo display logic.
- Removed redundant local imports of shouldDisplayChainLogo to streamline the codebase.
- Enhanced code organization by centralizing the logo display logic in a dedicated utility file.
