+++ clients/desktop/src/coin/ui/inputs/ChainOption.tsx
@@ -1,5 +1,6 @@
 import { Coin } from '@core/chain/coin/Coin'
 import { isFeeCoin } from '@core/chain/coin/utils/isFeeCoin'
+import { shouldDisplayChainLogo } from '@core/ui/chain/utils/shouldDisplayChainLogo'
 import { CheckIcon } from '@lib/ui/icons/CheckIcon'
 import { HStack, VStack } from '@lib/ui/layout/Stack'
 import { panel } from '@lib/ui/panel/Panel'
@@ -10,7 +11,6 @@ import styled from 'styled-components'
 
 import { ChainCoinIcon } from '../../../chain/ui/ChainCoinIcon'
 import { getChainEntityIconSrc } from '../../../chain/utils/getChainEntityIconSrc'
-import { shouldDisplayChainLogo } from '../../../vault/chain/utils'
 import { useSide } from '../../../vault/swap/providers/SideProvider'
 import { useFromCoin } from '../../../vault/swap/state/fromCoin'
 import { useToCoin } from '../../../vault/swap/state/toCoin'
