+++ clients/desktop/src/coin/ui/inputs/CoinOption.tsx
@@ -2,6 +2,7 @@ import { fromChainAmount } from '@core/chain/amount/fromChainAmount'
 import { extractAccountCoinKey } from '@core/chain/coin/AccountCoin'
 import { Coin } from '@core/chain/coin/Coin'
 import { isFeeCoin } from '@core/chain/coin/utils/isFeeCoin'
+import { shouldDisplayChainLogo } from '@core/ui/chain/utils/shouldDisplayChainLogo'
 import { HStack, VStack } from '@lib/ui/layout/Stack'
 import { panel } from '@lib/ui/panel/Panel'
 import { IsActiveProp, OnClickProp, ValueProp } from '@lib/ui/props'
@@ -15,7 +16,6 @@ import { ChainCoinIcon } from '../../../chain/ui/ChainCoinIcon'
 import { useFormatFiatAmount } from '../../../chain/ui/hooks/useFormatFiatAmount'
 import { getChainEntityIconSrc } from '../../../chain/utils/getChainEntityIconSrc'
 import { Skeleton } from '../../../components/skeleton'
-import { shouldDisplayChainLogo } from '../../../vault/chain/utils'
 import { useCurrentVaultCoin } from '../../../vault/state/currentVaultCoins'
 import { getCoinLogoSrc } from '../../logo/getCoinLogoSrc'
 import { useBalanceQuery } from '../../query/useBalanceQuery'
