+++ clients/desktop/src/coin/ui/inputs/SwapCoinInputField/index.tsx
@@ -1,5 +1,6 @@
 import { Coin } from '@core/chain/coin/Coin'
 import { isFeeCoin } from '@core/chain/coin/utils/isFeeCoin'
+import { shouldDisplayChainLogo } from '@core/ui/chain/utils/shouldDisplayChainLogo'
 import { ChevronDownIcon } from '@lib/ui/icons/ChevronDownIcon'
 import { ChevronRightIcon } from '@lib/ui/icons/ChevronRightIcon'
 import { HStack } from '@lib/ui/layout/Stack'
@@ -11,7 +12,6 @@ import { useTranslation } from 'react-i18next'
 import { ChainCoinIcon } from '../../../../chain/ui/ChainCoinIcon'
 import { ChainEntityIcon } from '../../../../chain/ui/ChainEntityIcon'
 import { getChainEntityIconSrc } from '../../../../chain/utils/getChainEntityIconSrc'
-import { shouldDisplayChainLogo } from '../../../../vault/chain/utils'
 import { ManageFromAmount } from '../../../../vault/swap/form/amount/ManageFromAmount'
 import { ToAmount } from '../../../../vault/swap/form/amount/ToAmount'
 import { SwapCoinBalance } from '../../../../vault/swap/form/SwapCoinBalance'
