+++ clients/desktop/src/vault/swap/form/SwapCoinInput.tsx
@@ -2,6 +2,7 @@ import { CoinKey } from '@core/chain/coin/Coin'
 import { isFeeCoin } from '@core/chain/coin/utils/isFeeCoin'
 import { isNativeCoin } from '@core/chain/coin/utils/isNativeCoin'
 import { swapEnabledChains } from '@core/chain/swap/swapEnabledChains'
+import { shouldDisplayChainLogo } from '@core/ui/chain/utils/shouldDisplayChainLogo'
 import { Opener } from '@lib/ui/base/Opener'
 import { ChevronDownIcon } from '@lib/ui/icons/ChevronDownIcon'
 import { HStack } from '@lib/ui/layout/Stack'
@@ -20,7 +21,6 @@ import { ChainOption } from '../../../coin/ui/inputs/ChainOption'
 import { CoinOption } from '../../../coin/ui/inputs/CoinOption'
 import { SelectItemModal } from '../../../coin/ui/inputs/SelectItemModal'
 import { SwapCoinInputField } from '../../../coin/ui/inputs/SwapCoinInputField'
-import { shouldDisplayChainLogo } from '../../chain/utils'
 import {
   useCurrentVaultCoin,
   useCurrentVaultCoins,
