+++ core/ui/chain/utils/shouldDisplayChainLogo.ts
